code-library/misc-2/requirements.txt
# requirements.txt
# Core dependencies for the fiber optic analysis system

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0

# Data processing
numpy>=1.21.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Configuration and utilities
PyYAML>=6.0
python-box>=6.0.0

# Data augmentation (optional but recommended)
albumentations>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web interface (optional)
gradio>=3.0.0

# Distributed training
nvidia-ml-py3

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Metrics and evaluation
scikit-learn>=1.1.0
